Refactor: Moved ZoneStage to components folder and updated imports
